Fixed property source order in spring env
* tsc4j loaded configs are now being preferred over yaml/properties
  configurations
* `Tsc4jPropertySource` is now initialized very early during spring boot
  startup.
* removed redundant classes
* moved tests in a different package to check if all settings do
  actually work.
