--- tsc4j-spring/src/main/java/com/github/tsc4j/spring/EnableTsc4j.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2017 - 2021 tsc4j project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.tsc4j.spring;
-
-import com.github.tsc4j.api.ReloadableConfig;
-import org.springframework.context.annotation.Import;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Enables {@link ReloadableConfig} as spring configuration source
- *
- * @see ReloadableConfig
- * @see Tsc4jSpringConfiguration
- */
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Import({Tsc4jSpringConfiguration.class})
-public @interface EnableTsc4j {
-}
