+++ tsc4j-spring/src/main/java/com/github/tsc4j/spring/Tsc4jBootstrapApplicationListener.java
@@ -16,6 +16,7 @@
 
 package com.github.tsc4j.spring;
 
+import com.github.tsc4j.core.Tsc4jImplUtils;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
@@ -24,8 +25,13 @@ import org.springframework.context.ApplicationListener;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.core.Ordered;
 import org.springframework.core.env.ConfigurableEnvironment;
+import org.springframework.core.env.MapPropertySource;
+import org.springframework.core.env.MutablePropertySources;
+import org.springframework.core.env.PropertySource;
 
+import java.util.Optional;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.function.Predicate;
 
 /**
  * Tsc4j spring early-startup application listener which initializes {@link com.github.tsc4j.api.ReloadableConfig} and
@@ -40,25 +46,15 @@ public final class Tsc4jBootstrapApplicationListener implements Ordered, Applica
         return Integer.MAX_VALUE;
     }
 
-    /**
-     * Tells whether given spring environment is part of spring app boostrap.
-     *
-     * @param env spring env
-     * @return true/false
-     */
-    private boolean isSpringAppBootstrap(ConfigurableEnvironment env) {
-        val cfgName = env.getProperty("spring.config.name", "").trim();
-        return cfgName.equalsIgnoreCase("bootstrap");
-    }
-
     @Override
     public void onApplicationEvent(@NonNull ApplicationPreparedEvent event) {
-        if (hasRegistrationAlreadyHappened()) {
-            return;
-        }
-
         val appCtx = event.getApplicationContext();
         val env = appCtx.getEnvironment();
+
+        if (hasRegistrationAlreadyHappened()) {
+            makeSureThatTsc4jPropSourceIsPreferred(env);
+            return;
+        }
 
         if (isBootstrapDebugEnabled(env)) {
             log.info("{}", SpringUtils.debugPropertySources(env));
@@ -71,6 +67,20 @@ public final class Tsc4jBootstrapApplicationListener implements Ordered, Applica
         }
     }
 
+    private Optional<PropertySource<?>> findPropertySource(MutablePropertySources mps,
+                                                           Predicate<PropertySource<?>> predicate) {
+        return mps.stream()
+            .filter(predicate::test)
+            .findFirst();
+    }
+
+    /**
+     * Tells whether tsc4j property should be registered
+     *
+     * @param appCtx application context
+     * @param env    spring environment
+     * @return true/false
+     */
     private boolean shouldRegisterPropertySource(ConfigurableApplicationContext appCtx,
                                                  ConfigurableEnvironment env) {
         if (!isTsc4jBootstrapEnabled(env)) {
@@ -89,14 +99,12 @@ public final class Tsc4jBootstrapApplicationListener implements Ordered, Applica
      */
     private void registerTsc4jPropertySource(ConfigurableEnvironment env) {
         log.info("initializing tsc4j-spring");
-        log.debug("trying to register tsc4j-spring property source to spring environment: {}", env);
 
         val rc = SpringUtils.reloadableConfig(env);
         val propSource = new Tsc4jPropertySource(rc);
         log.debug("created property source: {}" + propSource);
 
-        //env.getPropertySources().addLast(propSource);
-        env.getPropertySources().addFirst(propSource);
+        env.getPropertySources().addLast(propSource);
         log.debug("successfully registered tsc4j-spring property source.");
     }
 
@@ -109,6 +117,42 @@ public final class Tsc4jBootstrapApplicationListener implements Ordered, Applica
 
     private static boolean hasRegistrationAlreadyHappened() {
         return !registrationAlreadyDone.compareAndSet(false, true);
+    }
+
+    /**
+     * Makes sure that tsc4j property source is being placed before normal spring app's property source.
+     *
+     * @param env spring env
+     */
+    private void makeSureThatTsc4jPropSourceIsPreferred(ConfigurableEnvironment env) {
+        val mps = env.getPropertySources();
+
+        // find tsc4j prop source
+        val tsc4PropSource = findPropertySource(mps, this::isTsc4jPropertySource);
+        if (!tsc4PropSource.isPresent()) {
+            return;
+        }
+
+        // find normal spring config prop source
+        findPropertySource(mps, this::isSpringPropSource)
+            .map(PropertySource::getName)
+            .ifPresent(springPsName -> {
+                // remove tsc4j property source, so that we can re-add it later
+                val tsc4jPs = mps.remove(tsc4PropSource.get().getName());
+
+                // put it before first spring's property source
+                mps.addBefore(springPsName, tsc4jPs);
+
+                log.trace("reordered property sources: {}", SpringUtils.debugPropertySources(env));
+            });
+    }
+
+    private boolean isSpringPropSource(PropertySource<?> ps) {
+        return ps instanceof MapPropertySource && ps.getName().startsWith("applicationConfig:");
+    }
+
+    private boolean isTsc4jPropertySource(PropertySource<?> ps) {
+        return ps instanceof Tsc4jPropertySource && ps.getName().equals(Tsc4jImplUtils.NAME);
     }
 
     /**
