--- tsc4j-spring/src/main/java/com/github/tsc4j/spring/Tsc4jContextRefreshSpringConfiguration.java
@@ -1,54 +1,0 @@
-/*
- * Copyright 2017 - 2021 tsc4j project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.tsc4j.spring;
-
-import com.github.tsc4j.api.ReloadableConfig;
-import lombok.NonNull;
-import lombok.extern.slf4j.Slf4j;
-import lombok.val;
-import org.springframework.beans.factory.config.ConfigurableBeanFactory;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.cloud.context.refresh.ContextRefresher;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Scope;
-
-@Slf4j
-@Configuration
-public class Tsc4jContextRefreshSpringConfiguration {
-    /**
-     * Creates spring context refresher that refreshes spring-context when tsc4j configuration changes
-     *
-     * @param contextRefresher context refresher
-     * @param reloadableConfig reloadable config
-     * @return spring context refresher
-     * @see Constants#PROPERTY_REFRESH_CONTEXT
-     */
-    @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_SINGLETON)
-    @ConditionalOnMissingBean
-    @ConditionalOnClass(ContextRefresher.class)
-    @ConditionalOnProperty(value = Constants.PROPERTY_REFRESH_CONTEXT, matchIfMissing = true)
-    public Tsc4jSpringContextRefresher tsc4jSpringContextRefresher(@NonNull ContextRefresher contextRefresher,
-                                                                   @NonNull ReloadableConfig reloadableConfig) {
-        val refresher = new Tsc4jSpringContextRefresher(contextRefresher, reloadableConfig);
-        log.debug("supplying spring-context refresher singleton: {}", refresher);
-        return refresher;
-    }
-}
