--- tsc4j-spring/src/main/java/com/github/tsc4j/spring/Tsc4jHealthIndicatorSpringConfiguration.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2017 - 2021 tsc4j project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.tsc4j.spring;
-
-import com.github.tsc4j.api.ReloadableConfig;
-import lombok.NonNull;
-import lombok.extern.slf4j.Slf4j;
-import lombok.val;
-import org.springframework.beans.factory.config.ConfigurableBeanFactory;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Scope;
-
-@Slf4j
-@Configuration
-public class Tsc4jHealthIndicatorSpringConfiguration {
-    /**
-     * Creates tsc4j health indicator.
-     *
-     * @param reloadableConfig reloadable config
-     * @return health indicator
-     */
-    @Bean
-    @Scope(ConfigurableBeanFactory.SCOPE_SINGLETON)
-    @ConditionalOnMissingBean
-    @ConditionalOnClass(name = "org.springframework.boot.actuate.health.HealthIndicator")
-    @ConditionalOnProperty(value = Constants.PROPERTY_ENABLED, matchIfMissing = true)
-    Tsc4jHealthIndicator tsc4jHealthIndicator(@NonNull ReloadableConfig reloadableConfig) {
-        val indicator = new Tsc4jHealthIndicator(reloadableConfig);
-        log.debug("supplying tsc4j health indicator singleton: {}", indicator);
-        return indicator;
-    }
-}
