+++ tsc4j-spring/src/test/groovy/com/github/tsc4j/spring/Tsc4jPropertySourceSpec.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 - 2019 tsc4j project
+ * Copyright 2017 - 2021 tsc4j project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,12 +12,12 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
 package com.github.tsc4j.spring
 
 import com.github.tsc4j.api.ReloadableConfig
+import com.github.tsc4j.spring.Tsc4jPropertySource
 import com.github.tsc4j.test.TestReloadable
 import com.typesafe.config.ConfigFactory
 import spock.lang.Specification
