+++ tsc4j-spring/src/test/resources/application-test.conf
@@ -32,5 +32,11 @@ hocon {
   }
 }
 
+# the following section is defined in HOCON and YAML; values from HOCON should be preferred
+common.settings {
+  str: value-from-hocon
+  list: [ 5, 6, 7 ]
+}
+
 # vim:syntax=hocon shiftwidth=2 softtabstop=2 expandtab
 # EOF
